Normalize registry operation namespaces
