+++ server/modules/providers/database/mssql_provider/registry.py
@@ -1086,6 +1086,7 @@ def _security_roles_remove_member(args: Dict[str, Any]):
   return (DbRunMode.EXEC, sql, (role, user_guid, user_guid))
 
 @register("db:system:assistant_personas:get_by_name:1")
+@register("db:system:personas:get_by_name:1")
 def _assistant_personas_get_by_name(args: Dict[str, Any]):
   name = args["name"]
   sql = """
@@ -1125,6 +1126,7 @@ def _assistant_models_list(_: Dict[str, Any]):
   return (DbRunMode.JSON_MANY, sql, ())
 
 @register("db:system:assistant_models:list:1")
+@register("db:system:models:list:1")
 def _db_system_assistant_models_list(args: Dict[str, Any]):
   return _assistant_models_list(args)
 
@@ -1155,6 +1157,7 @@ def _assistant_personas_list(_: Dict[str, Any]):
   return (DbRunMode.JSON_MANY, sql, ())
 
 @register("db:system:assistant_personas:list:1")
+@register("db:system:personas:list:1")
 def _db_system_assistant_personas_list(args: Dict[str, Any]):
   return _assistant_personas_list(args)
 
@@ -1206,6 +1209,7 @@ async def _assistant_personas_upsert(args: Dict[str, Any]):
   )
 
 @register("db:system:assistant_personas:upsert:1")
+@register("db:system:personas:upsert:1")
 async def _db_system_assistant_personas_upsert(args: Dict[str, Any]):
   return await _assistant_personas_upsert(args)
 
@@ -1224,10 +1228,12 @@ def _assistant_personas_delete(args: Dict[str, Any]):
   return (DbRunMode.EXEC, sql, params)
 
 @register("db:system:assistant_personas:delete:1")
+@register("db:system:personas:delete:1")
 def _db_system_assistant_personas_delete(args: Dict[str, Any]):
   return _assistant_personas_delete(args)
 
 @register("db:system:assistant_models:get_by_name:1")
+@register("db:system:models:get_by_name:1")
 def _assistant_models_get_by_name(args: Dict[str, Any]):
   name = args["name"]
   sql = """
@@ -1239,6 +1245,7 @@ def _assistant_models_get_by_name(args: Dict[str, Any]):
   return (DbRunMode.JSON_ONE, sql, (name,))
 
 @register("db:system:assistant_conversations:insert:1")
+@register("db:system:conversations:insert:1")
 def _assistant_conversations_insert(args: Dict[str, Any]):
   personas_recid = args["personas_recid"]
   models_recid = args["models_recid"]
@@ -1278,6 +1285,7 @@ def _assistant_conversations_insert(args: Dict[str, Any]):
   )
 
 @register("db:system:assistant_conversations:find_recent:1")
+@register("db:system:conversations:find_recent:1")
 def _assistant_conversations_find_recent(args: Dict[str, Any]):
   personas_recid = args["personas_recid"]
   models_recid = args["models_recid"]
@@ -1325,6 +1333,7 @@ def _assistant_conversations_find_recent(args: Dict[str, Any]):
   )
 
 @register("db:system:assistant_conversations:update_output:1")
+@register("db:system:conversations:update_output:1")
 def _assistant_conversations_update_output(args: Dict[str, Any]):
   recid = args["recid"]
   output_data = args.get("output_data")
@@ -1338,6 +1347,7 @@ def _assistant_conversations_update_output(args: Dict[str, Any]):
   return (DbRunMode.EXEC, sql, (output_data, tokens, recid))
 
 @register("db:system:assistant_conversations:list_by_time:1")
+@register("db:system:conversations:list_by_time:1")
 def _assistant_conversations_list_by_time(args: Dict[str, Any]):
   personas_recid = args["personas_recid"]
   start = args["start"]
@@ -1362,6 +1372,7 @@ def _assistant_conversations_list_by_time(args: Dict[str, Any]):
 
 
 @register("db:system:assistant_conversations:list_recent:1")
+@register("db:system:conversations:list_recent:1")
 def _assistant_conversations_list_recent(_: Dict[str, Any]):
   sql = """
     SELECT TOP (2)
