+++ server/registry/system/__init__.py
@@ -27,11 +27,11 @@ __all__ = [
 def register(router: "RegistryRouter") -> None:
   domain = router.domain("system")
   config.register(domain.subdomain("config"))
-  conversations.register(domain.subdomain("conversations", op_segment="assistant_conversations"))
+  conversations.register(domain.subdomain("conversations"))
   guilds.register(domain.subdomain("guilds"))
   links.register(domain.subdomain("links"))
-  models.register(domain.subdomain("models", op_segment="assistant_models"))
-  personas.register(domain.subdomain("personas", op_segment="assistant_personas"))
+  models.register(domain.subdomain("models"))
+  personas.register(domain.subdomain("personas"))
   roles.register(domain.subdomain("roles"))
   routes.register(domain.subdomain("routes"))
   public_users.register(domain.subdomain("public_users"))
