+++ server/registry/users/content/__init__.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from . import cache, files, profile, public
+from server.registry.users import cache, files, profile, public
 
 if TYPE_CHECKING:
   from server.registry import DomainRouter
@@ -19,8 +19,7 @@ __all__ = [
 
 
 def register(domain: "DomainRouter") -> None:
-  content_router = domain.subdomain("content", op_segment="")
-  cache.register(content_router.subdomain("cache", op_segment="content_cache"))
-  files.register(content_router.subdomain("files", op_segment="content_files"))
-  public.register(content_router.subdomain("public", op_segment="content_public"))
-  profile.register(content_router.subdomain("profile", op_segment="content_profile"))
+  cache.register(domain.subdomain("cache"))
+  files.register(domain.subdomain("files"))
+  public.register(domain.subdomain("public"))
+  profile.register(domain.subdomain("profile"))
