+++ server/registry/users/files/__init__.py
@@ -1,0 +1,30 @@
+"""Users files registry bindings."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from server.registry.types import DBRequest
+
+if TYPE_CHECKING:
+  from server.registry import SubdomainRouter
+
+__all__ = [
+  "set_gallery_request",
+  "register",
+]
+
+
+def set_gallery_request(user_guid: str, name: str, gallery: bool) -> DBRequest:
+  return DBRequest(
+    op="db:users:files:set_gallery:1",
+    params={
+      "user_guid": user_guid,
+      "name": name,
+      "gallery": bool(gallery),
+    },
+  )
+
+
+def register(router: "SubdomainRouter") -> None:
+  router.add_function("set_gallery", version=1)
